app/requirements.txt
# ─── Core ─────────────────────────────────────────────────────────────────────
streamlit>=1.35.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
pyarrow>=14.0.0          # Parquet support

# ─── Machine Learning ─────────────────────────────────────────────────────────
scikit-learn>=1.3.0
shap>=0.44.0             # XAI (optional but recommended)

# ─── NLP & Embeddings ─────────────────────────────────────────────────────────
gensim>=4.3.0            # Word2Vec
transformers>=4.40.0     # BERT, DistilBART, RoBERTa, Helsinki-NLP
torch>=2.0.0             # PyTorch backend for HuggingFace
sentencepiece>=0.1.99    # Required for Helsinki-NLP translation
sacremoses>=0.1.1        # Tokenizer for translation model

# ─── RAG / LLM ────────────────────────────────────────────────────────────────
openai>=1.10.0           # OpenAI-compatible client (works with Ollama too)
google-generativeai>=0.7.0
# ─── Visualization ────────────────────────────────────────────────────────────
plotly>=5.18.0           # Interactive charts in Streamlit

# ─── Optional (used in the original notebooks, not strictly required by app) ──
# faiss-cpu>=1.7.4       # FAISS index for ANN search (uncomment if using FAISS)
# tensorflow>=2.13.0     # Only if loading Keras CNN models directly
# rouge-score>=0.1.2     # For translation evaluation display
